backend/requirements.txt
asgiref==3.8.1
bleach==6.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.4
cryptography==43.0.1
cssbeautifier==1.15.1
Django==5.1.1
django-allauth==65.0.1
django-cors-headers==4.6.0
django-resized==1.0.2
django-summernote==0.8.20.0
djangorestframework==3.15.2
djlint==1.35.2
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.10
iniconfig==2.0.0
jsbeautifier==1.15.1
json5==0.9.25
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
urllib3==2.2.3
webencodings==0.5.1
